Fix load drivers on module import (now uses 'lazy' loading)
When loading the drivers right on import, some problem with cache could
cause a stack overflow (recursive import, probably). So now I’m
changing it to only load the drivers when the user calls a function in
Loki. That way, by the time the driver is loaded, Loki is already fully
imported and no recursive import will take place.